requirements.txt
# Core dependencies
numpy>=1.22.0
pandas>=1.5.0
matplotlib>=3.5.0
tqdm>=4.65.0
PyYAML>=6.0.0
requests>=2.31.0

# Geospatial dependencies
rasterio>=1.3.0 